add log function to helpers
